+++ CHANGELOG.md
@@ -4,16 +4,18 @@
 * renamed `Poolable` to `Pooled`
 * renamed `ComponentPoolable` to `PooledComponent`
 * removed `FastMath` and `Utils`, unrelated to ECS
+* `time` and `frame` getters have been moved from `World` to `EntitySystem`, `World` has methods instead
 ### API Changes
 * deprecated `ComponentMapper` use `Mapper` instead
 * deprecated `ComponentMapper#get(Entity)`, use `Mapper[Entity]` instead
+* properties have been added to the `World`, can be accessed using the `[]` operator
+* `System`s can be assigned to a group when adding them to the `World`, `Word.process()` can be called for a specific group
 ### Enhancements
 * created a code generating transformer to insert `Manager`, `System`s and `Mapper`s into `Manager` and `System` instances
-* properties have been added to the world, accessible using the `[]` operator 
 ### Bugfixes
 * DelayedEntityProcessingSystem keeps running until all current entities have expired
 ### Internal
-* upgraded unittest dependency
+* upgraded dependencies
 
 ##0.6.0
 ### API Changes